fstab.twrp: Don't touch firmware (apnhlos) and firmware-modem (modem)
Those two having problem when restoring from a backup so result in a
corrupted files.

I already try to backup and restore it as "emmc" but result still
the same.

Beside.. these partition are mounted as RO on Android system so not
likely get corrupted by user or anything so we shouldn't touch it.